Use a stack size slightly smaller than 1MB
This is motivated by discussions in https://github.com/status-im/nim-beacon-chain/pull/924

The stack size on Android and iOS is 1MB, so we want to detect
any possible stack overflow issues there while we are testing
our software on the desktop. The limit is 1 million bytes instead
of exactly 1MB, so we can be on the safe side.