Work around sync not dealing with gaps
Currently, when a peer returns fewer slots than we expected, the next
"chunk" of blocks will nonetheless be requested as if the full request
was satisfied. This results in an infinite loop where blocks without
parent are passed to be block pool and rejected.

The workaround here is to request one range at a time, extending the
range every time - this way, we'll request the same range as the
previous request, in addition to what should have been requested.