discovery: don't filter peers during sync
We don't care about peer attestation subnet subscriptions during block sync.

The ugly bit here is informing that part of the code that calls
`queryRandom()` of the block sync status.

TODO: keep a percentage of unfiltered peers in our peer pool so we can
discover the mesh layout quicker and more reliably. What's useless for
attestation propagation can be useful for discovery.