update to latest light client libp2p protocol
Incorporates the latest changes to the light client sync protocol based
on Devconnect AMS feedback. Note that this breaks compatibility with the
previous prototype, due to changes to data structures and endpoints.
See https://github.com/ethereum/consensus-specs/pull/2802

Due to the amount of changes, it is best to review the heavily edited
files from scratch (likewise, for the spec).